pyproject.toml
[project]
name = "crashtransformer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "bert-score>=0.3.13",
    "datasets>=4.2.0",
    "evaluate>=0.4.6",
    "grpcio==1.60.1",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0,<1.0.0",
    "langchain-google-genai>=2.0.0",
    "langchain-groq>=0.2.0",
    "langchain-openai>=0.2.0",
    "markdown>=3.9",
    "neo4j>=6.0.2",
    "nltk>=3.9.2",
    "ollama>=0.5.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pygments>=2.19.2",
    "pymdown-extensions>=10.16.1",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.14.1",
    "rouge-score>=0.1.2",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.1",
    "torch>=2.7.1",
    "transformers>=4.54.0",
]
